+++ R/tar_script.R
@@ -31,7 +31,7 @@
 #'       a pipeline object, which usually means ending with a call to
 #'       [list()]. In practice, (3) and (4) can be combined together
 #'       in the same function call.
-#' @return Nothing.
+#' @return `NULL` (invisibly).
 #' @param code R code to write to `_targets.R`. If `NULL`, an example
 #'   target script is written instead.
 #' @param library_targets logical, whether to write a `library(targets)`
@@ -43,6 +43,10 @@
 #'   (Set to `"true"` or `"false"` with `Sys.setenv()`).
 #'   If `ask` and the `TAR_ASK` environment variable are both
 #'   indeterminate, defaults to `interactive()`.
+#' @param tidy_eval Logical, whether to enable tidy evaluation
+#'   to customize the script output. Disabled by default
+#'   because [tar_target()] also has a `tidy_eval` argument.
+#' @param envir Environment for tidy evaluation.
 #' @examples
 #' tar_dir({ # tar_dir() runs code from a temporary directory.
 #' tar_script() # Writes an example target script.
@@ -53,15 +57,32 @@
 #'   list(x)
 #' }, ask = FALSE)
 #' writeLines(readLines("_targets.R"))
+#' # With tidy eval:
+#' name <- quote(my_target)
+#' tar_script({
+#'   x <- tar_target(!!name, 1 + 1)
+#'   tar_option_set()
+#'   list(x)
+#' }, ask = FALSE, tidy_eval = TRUE)
+#' writeLines(readLines("_targets.R"))
 #' })
-tar_script <- function(code = NULL, library_targets = TRUE, ask = NULL) {
+tar_script <- function(
+  code = NULL,
+  library_targets = TRUE,
+  ask = NULL,
+  tidy_eval = FALSE,
+  envir = parent.frame()
+) {
   if (!tar_should_overwrite(ask, path_script())) {
     # covered in tests/interactive/test-tar_script.R # nolint
     return(invisible()) # nocov
   }
+  force(envir)
   assert_lgl(library_targets, "library_targets must be logical.")
   assert_scalar(library_targets, "library_targets must have length 1.")
-  code <- substitute(code)
+  assert_lgl(tidy_eval, "tidy_eval must be logical.")
+  assert_envir(envir, "envir must be an environment.")
+  code <- tidy_eval(substitute(code), envir, tidy_eval)
   text <- trn(
     length(code),
     parse_target_script_code(code),